requirements.txt
# Python依赖

# 核心依赖
requests>=2.31.0
pyyaml>=6.0
sqlparse>=0.4.4

# 数据库连接
mysql-connector-python>=8.0.33

# JSON处理
simplejson>=3.19.1

# CLI工具
click>=8.1.3
rich>=13.3.5

# 测试
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# 代码质量
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# 可选：如果使用本地模型
# transformers>=4.30.0
# torch>=2.0.0
# accelerate>=0.20.0

# 可选：如果需要更高级的SQL解析
# sqlglot>=11.0.0
# pglast>=5.0

# 开发工具
ipython>=8.12.0
jupyter>=1.0.0
